Search Bar: The new category works for any UIViewController. Renamed it accordingly
